+++ app/controllers/department.controller.js
@@ -1,70 +1,77 @@
 const geoCacheService = require('services/geoCacheService');
+const DepartmentService = require('services/departmentService');
 const Department = require('models/department.model');
 
-exports.findAll = async (request, response) => {
+exports.getDepartments = async (req, res) => {
   try {
-    const data = await Department.getAll(request.sorting);
+    const { page, limit, sortField, sortOrder } = req.processedQuery;
+    const options = {
+      page,
+      limit,
+      sort: { field: sortField, order: sortOrder },
+    };
+    const data = await DepartmentService.findAll(options);
 
-    if (!data) {
-      return response.status(404).send({ message: 'Departments not found' });
+    if (!data || data.length === 0) {
+      return res.status(404).send({ message: 'Departments not found' });
     }
 
-    return response.status(200).json({ data });
+    return res.status(200).json({ data, page, limit });
   } catch (error) {
-    return response.status(403).send({
-      message: request.polyglot.t('failed_to_retrieve_departments') || error.message,
+    return res.status(403).send({
+      message: req.polyglot.t('failed_to_retrieve_departments') || error.message,
     });
   }
 };
 
-exports.findByLngLat = async (request, response) => {
+exports.findByLngLat = async (req, res) => {
   try {
-    const { lng, lat } = request.params;
+    const { lng, lat } = req.params;
 
     if (!lng || !lat) {
-      return response.status(400).send({ message: 'Longitude and latitude are required' });
+      return res.status(400).send({ message: 'Longitude and latitude are required' });
     }
 
     // Check if the data is already cached.
     const cachedData = await geoCacheService.getCachedLocation(lng, lat);
     if (cachedData) {
-      return response.status(200).json(cachedData);
+      return res.status(200).json(cachedData);
     }
 
     const data = await Department.findByLngLat(lng, lat);
 
     if (!data) {
-      return response.status(404).send({ message: 'Departments not found' });
+      return res.status(404).send({ message: 'Departments not found' });
     }
 
     // Update cache.
     geoCacheService.cacheLocation(lng, lat, data).catch((error) => console.error('Error: ', error));
 
-    return response.status(200).json(data);
+    return res.status(200).json(data);
   } catch (error) {
-    return response.status(403).send({
-      message: request.polyglot.t('failed_to_retrieve_department') || error.message,
+    return res.status(403).send({
+      message: req.polyglot.t('failed_to_retrieve_department') || error.message,
     });
   }
 };
 
-exports.findById = async (request, response) => {
+exports.getDepartmentById = async (req, res) => {
   try {
-    const { id } = request.params;
+    const { id } = req.params;
 
     if (!id) {
-      return response.status(400).send({ message: 'ID is required' });
+      return res.status(400).send({ message: 'ID is required' });
     }
 
     const data = await Department.findById(id);
 
     if (!data) {
-      return response.status(404).send({ message: 'Department not found' });
+      return res.status(404).send({ message: 'Department not found' });
     }
-    return response.status(200).json(data);
+    return res.status(200).json(data);
   } catch (error) {
-    return response.status(403).send({
-      message: request.polyglot.t('failed_to_retrieve_department') || error.message,
+    return res.status(403).send({
+      message: req.polyglot.t('failed_to_retrieve_department') || error.message,
     });
   }
 };